done with base calculation testing
